move handleAutoplay to separate file
