+++ packages/playback-core/src/index.ts
@@ -1,6 +1,7 @@
 import mux, { Options } from "mux-embed";
 
 import Hls, { HlsConfig } from "hls.js";
+import { AutoplayTypes, handleAutoplay } from "./autoplay";
 import { isKeyOf } from "./util";
 export type ValueOf<T> = T[keyof T];
 
@@ -19,17 +20,6 @@ export const StreamTypes: StreamTypes = {
   VOD: "on-demand",
   LIVE: "live",
   LL_LIVE: "ll-live",
-};
-
-// TODO add INVIEW_MUTED, INVIEW_ANY
-export type AutoplayTypes = {
-  ANY: "any";
-  MUTED: "muted";
-};
-
-export const AutoplayTypes: AutoplayTypes = {
-  ANY: "any",
-  MUTED: "muted",
 };
 
 export type ExtensionMimeTypeMap = {
@@ -405,33 +395,7 @@ export const loadMedia = (
       });
     } else if (props.autoplay) {
       hls.once(Hls.Events.MANIFEST_PARSED, () => {
-        const oldMuted = mediaEl.muted;
-        const restoreMuted = () => (mediaEl.muted = oldMuted);
-        switch (props.autoplay) {
-          // ANY:
-          // try to play with current options
-          // if it fails, mute and try playing again
-          // if that fails, restore muted state and don't try playing again
-          case AutoplayTypes.ANY:
-            mediaEl.play().catch((error: Error) => {
-              mediaEl.muted = true;
-              mediaEl.play().catch(restoreMuted);
-            });
-            break;
-          // MUTED:
-          // mute the player and then try playing
-          // if that fails, restore muted state
-          case AutoplayTypes.MUTED:
-            mediaEl.muted = true;
-            mediaEl.play().catch(restoreMuted);
-
-            break;
-          // Default or if autoplay is a boolean attribute:
-          // Try playing the video and catch the failed autoplay warning
-          default:
-            mediaEl.play().catch(() => {});
-            break;
-        }
+        handleAutoplay(mediaEl, props.autoplay);
       });
     }
